add weightscore roadmap and remove warning

+++ src/main/java/com/exe/skillverse_backend/ai_service/dto/response/RoadmapSessionSummary.java
@@ -30,7 +30,7 @@ public class RoadmapSessionSummary {
     // Progress tracking
     private Integer totalQuests;
     private Integer completedQuests;
-    private Integer progressPercentage;
+    private Double progressPercentage;
 
     // Metadata
     private String difficultyLevel;
